server/requirements.txt
attrs==21.2.0
boto3==1.21.22
botocore==1.15.0
certifi==2021.5.30
cffi==1.15.0
chardet==3.0.4
click==8.0.3
cryptography==3.4.8
dnspython==2.0.0
docutils==0.15.2
future==0.18.2
idna==2.8
iniconfig==1.1.1
jmespath==0.10.0
marshmallow==3.10.0
packaging==21.2
pluggy==0.13.1
py==1.10.0
pycparser==2.20
PyJWT==1.7.1
pymongo==3.11.0
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
requests==2.22.0
s3transfer==0.3.7
six==1.16.0
toml==0.10.2
tornado==6.1
urllib3==1.25.11

pip~=21.3.1
wheel~=0.37.1
setuptools~=60.8.2
websockets~=10.2
